Remove jest stuff from benches
