--- test/benchmark/jest.config.js
@@ -1,7 +1,0 @@
-const baseConfig = require('../../jest.config')
-
-module.exports = Object.assign({}, baseConfig, {
-  verbose: false,
-  testEnvironment: 'node',
-  testPathIgnorePatterns: []
-})
